Move around lua files for better modularization
