+++ runtime/lua/lsp/client.lua
@@ -1,12 +1,12 @@
 local json = require('lsp.json')
-local util = require('runtime.lua.builtin_util')
+local util = require('neovim.util')
 
 local protocol = require('runtime.lua.lsp.protocol')
 local message = require('runtime.lua.lsp.message')
 local lsp_doautocmd = require('runtime.lua.lsp.autocmds').lsp_doautocmd
 local get_callback_function = require('runtime.lua.lsp.callbacks').get_callback_function
 
-local log = require('runtime.lua.lsp.lsp').log
+local log = require('neovim.log')
 
 local read_state = {
   init = 0,
