--- runtime/lua/lsp/lsp.lua
@@ -1,7 +1,0 @@
-local lsp = {}
-
--- logging setup
-lsp.log = require('runtime.lua.log')
-lsp.log.prefix = '[LSP]'
-
-return lsp
